:truck: Rename repo from CV to weiji14.github.io
Changing the repository from a pure 'Curriculum Vitae' site to a proper academic personal website! Might as well right? Will have a dedicated CV page later on. Updated the README.md with more comprehensive local build instructions in case anyone wants to reuse parts of it. Had to put the source code in the 'source' branch, and publish to the 'master' branch because things are now published to a user page. Also normalized some urlpaths in the base.njk file according to https://www.11ty.dev/docs/filters/url, in case anyone wants to use a different pathPrefix.