:children_crossing: Set header link anchors
Have proper 'permalink' anchors for each of the headings on my website! This has been on my todo list for the longest time, as navigating from the links in my CV page to sections in other pages (in community/academic/professional) has been broken for ages. Need to remote the date part from the academic permalinks, decamelize the slugs, and set the permalink style to header link.